requirements.txt
crawlab-sdk>=0.6.0
selenium==4.2.0
scrapy==2.6.1